requirements.txt
discord.py>=1.5
gspread
Flask
setuptools>=40.3.0